Made tinyca basedir configurable
The location where tinyca stored its data was hardcoded
($HOME/.TinyCA).  This change uses the location given in the
environment variable TINYCA_BASEDIR.  In case this variable is
undefined, it uses the default (still $HOME/.TinyCA).  Similarly, the
default export directory can be configured through the environment
variable TINYCAEXPORTDIR.  This was reported as Debian #454380.